% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT600.HW1.SLM-package.R
\docType{package}
\name{STAT600.HW1.SLM-package}
\alias{STAT600.HW1.SLM}
\alias{STAT600.HW1.SLM-package}
\title{STAT600.HW1.SLM: Univariate OLS with Rcpp}
\description{
Fits a simple OLS linear regression with one numeric predictor and one numeric response.
}
\author{
\strong{Maintainer}: Wentao Guo \email{Wentao.Guo@colostate.edu}

}
