% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT600.HW1.SLM-package.R
\docType{package}
\name{STAT600.HW1.SLM-package}
\alias{STAT600.HW1.SLM}
\alias{STAT600.HW1.SLM-package}
\title{STAT600.HW1.SLM: Univariate OLS with Rcpp}
\description{
Fits a simple OLS linear regression with one numeric predictor and one numeric response.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WentaoGuo19/STAT600.HW1.SLM}
  \item Report bugs at \url{https://github.com/WentaoGuo19/STAT600.HW1.SLM/issues}
}

}
\author{
\strong{Maintainer}: Wentao Guo \email{Wentao.Guo@colostate.edu}

}
